requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.6.2
Flask-WTF==1.0.1
Flask-Caching==2.0.1
Flask-SocketIO==5.3.2
Flask-Cors==3.0.10
SQLAlchemy==1.4.46
Werkzeug==2.0.3
gevent==22.10.2
gevent-websocket==0.10.1
psycopg2==2.9.5
python-dotenv==0.21.1
scikit-learn==1.0.2
pandas==1.5.3
numpy==1.24.2
joblib==1.2.0
matplotlib==3.7.0
seaborn==0.12.2
scrypt==0.8.27
